Implement authorization model on db and use it in decksRoute

+++ server/routes/decks.js
@@ -1,46 +1,68 @@
 var express = require('express')
 var router = express.Router()
 const decksModel = require('../../db/models/decks');
+const authorization = require('../../db/models/authorization');
 const tryCatch = require('../utils/tryCatch');
 
-router.get('/', (req, res)=>{
+router.get('/', async (req, res)=>{
     //query string like ?tableId=123
     const {tableId} = req.query;
-    //if req.user && user owns table
+    const userId = req.user;
     tryCatch(async()=>{
-        //this is where the monster query goes
-        const {rows: decks} = await decksModel.getByTableId(tableId);
-        res.status(200).send(decks)
-    })
+        const authorized = await authorization.user.ownsTable(userId, tableId)
+        if(authorized){
+            //this is where the monster query goes
+            const {rows: decks} = await decksModel.getByTableId(tableId);
+            res.status(200).send(decks)
+        } else {
+            res.status(401).send({message: "Unathorized"})
+        }
+    }, res)
 })
 
 router.post('/', (req, res)=>{
     const deck = req.body;
-    //if req.user
-    tryCatch(async ()=>{
-        const result = await decksModel.post(deck)
-        res.status(200).send(result)
-    })
+    if(req.user){
+        tryCatch(async ()=>{
+            const result = await decksModel.post(deck)
+            res.status(200).send(result)
+        }, res)
+    } else {
+        res.status(401).send({message: "Unathorized"})
+    }
 })
 
-router.put('/:id', (req, res)=>{
+router.put('/:id', async (req, res)=>{
     const deck = req.body;
     const id = req.params.id
     deck.id=id
-    //if req.user && user owns deck's table
-    tryCatch(async ()=>{
-        let result = await decksModel.put(deck)
-        console.log(result)
-        res.status(200).send(deck)
-    })
+    const tableId = deck.table_id;
+    const userId = req.user;
+    tryCatch(async()=>{
+        const authorized = await authorization.user.ownsTable(userId, tableId)
+        if(authorized){
+            let result = await decksModel.put(deck)
+            res.status(200).send(result)
+        } else {
+            res.status(401).send({message: "Unathorized"})
+        }
+    }, res)
 })
 
 router.delete('/:id', async (req, res)=>{
     const id = req.params.id
-    //if req.user && user owns table of deck
-    const result = await decksModel.delete(id);
-    console.log(result)
-    res.status(200).send(`Deleted deck ${id}`);
+    const userId = req.user;
+    tryCatch(async ()=>{
+        const deck = await decksModel.get(id);
+        const authorized = await authorization.user.ownsTable(userId, deck.table_id)
+        if(authorized){
+            const result = await decksModel.delete(id);
+            console.log(result)
+            res.status(200).send(`Deleted deck ${id}`);
+        } else {
+            res.status(401).send({message: "Unathorized"})
+        }
+    }, res)
 })
 
 module.exports = router;
